\name{VariantFiltering-defunct}
\alias{VariantFiltering-defunct}
\alias{fetchKnownVariantsByID,MafDb-method}
\alias{fetchKnownVariantsByID}


\title{Defunct functions in package \sQuote{VariantFiltering}}

\description{
  These functions are defunct and no longer available.
}

\details{
  The following functions are defunct; use the replacement
  indicated below:

  \itemize{
    \item{fetchKnownVariantsByID(): \code{\link{snpid2maf}()}}
  }
}
\author{
  Robert Castelo <robert.castelo@upf.edu>
}
\keyword{internal}
\keyword{misc}
