\name{VariantFiltering-deprecated}
\alias{VariantFiltering-deprecated}
\alias{fetchKnownVariantsByID,MafDb-method}
\alias{fetchKnownVariantsByID}


\title{Deprecated functions in package \sQuote{VariantFiltering}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{VariantFiltering} only, and will be defunct at the next release.
}

\details{
  There are currently no deprecated functions.
  %% The following functions are deprecated and will be made defunct; use
  %% the replacement indicated below:
  %%
  %% \itemize{
  %%   \item{fetchKnownVariantsByID(): \code{\link{snpid2maf}()}}
  %% }
}
\author{
  Robert Castelo <robert.castelo@upf.edu>
}
\keyword{internal}
\keyword{misc}
